mgr/cephadm: Add GrafanaService
